Reorganize shorts
